requirements.txt
# Telegram相关
telethon>=1.34.0
python-telegram-bot>=20.7

# 异步支持
asyncio>=3.4.3
aiofiles>=23.2.1

# 任务调度
APScheduler>=3.10.4

# 加密和安全
cryptography>=41.0.0

# 配置文件处理
PyYAML>=6.0.1
python-dotenv>=1.0.0

# 系统监控
psutil>=5.9.6

# 数据库
sqlalchemy>=2.0.23
aiosqlite>=0.19.0

# 文件监控
watchdog>=3.0.0

# 终端颜色
colorama>=0.4.6

# HTTP请求
requests>=2.31.0

# Web框架
Flask>=2.3.0
Flask-SocketIO>=5.3.0

# WSGI服务器
eventlet>=0.33.0
gunicorn>=21.0.0

# Web工具
Jinja2>=3.1.0
Werkzeug>=2.3.0